Add Posts controller refactor, other updates
