+++ app/controllers/posts_controller.rb
@@ -16,26 +16,16 @@
 class PostsController < ApplicationController
 
   before_action :authenticate_user!, :except => ['index', 'create', 'up_vote']
-  before_action :verify_admin, :only => ['new', 'edit', 'update', 'destroy']
-  before_action :instantiate_tracker
-
-#  before_filter :fetch_counts, :only => 'create'
-  after_action :send_message, :only => 'create'
-#  after_filter :view_causes_vote, :only => 'index'
+  #before_action :instantiate_tracker
+  # after_action :send_message, :only => 'create'
 
   layout "clean", only: [:index]
-
 
   def index
     @topic = Topic.undeleted.ispublic.where(id: params[:topic_id]).first#.includes(:forum)
     if @topic
       @posts = @topic.posts.ispublic.active.all.chronologic.includes(:user)
       @post = @topic.posts.new
-
-      #@related = Topic.ispublic.by_popularity.front.tagged_with(@topic.tag_list)
-
-      #@feed_link = "<link rel='alternate' type='application/rss+xml' title='RSS' href='#{topic_posts_url(@topic)}.rss' />"
-
       @page_title = "#{@topic.name}"
       @title_tag = "#{AppSettings['settings.site_name']}: #{@page_title}"
       add_breadcrumb t(:community, default: "Community"), forums_path
@@ -52,34 +42,10 @@ class PostsController < ApplicationController
     end
   end
 
-  def show
-    @post = Post.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.rhtml
-    end
-  end
-
   def new
     @topic = Topic.find(params[:topic_id], :include => :forum)
     @post = Post.new
     @forums = Forum.all
-  end
-
-  def edit
-    @post = Post.where(id: params[:id]).first
-
-    respond_to do |format|
-      format.js
-    end
-  end
-
-  def cancel
-    @post = Post.where(id: params[:id]).first
-
-    respond_to do |format|
-      format.js
-    end
   end
 
   def create
@@ -92,35 +58,10 @@ class PostsController < ApplicationController
 
     respond_to do |format|
       if @post.save
-
         format.html {
           @posts = @topic.posts.ispublic.chronologic.active
-          redirect_to @topic.doc_id.nil? ? topic_posts_path(@topic) : doc_path(@topic.doc_id)
-
-        }
-        format.js {
-          if params[:from] == 'admin' #posted from admin side
-            fetch_counts
-
-            @posts = @topic.posts.chronologic
-            @admins = User.admins
-            #@post = Post.new
-            case @post.kind
-            when "reply"
-              @tracker.event(category: "Agent: #{current_user.name}", action: "Agent Replied", label: @topic.to_param) #TODO: Need minutes
-            when "note"
-              @tracker.event(category: "Agent: #{current_user.name}", action: "Agent Posted Note", label: @topic.to_param) #TODO: Need minutes
-            end
-            render 'admin/ticket'
-
-          else # posted from customer side
-            @posts = @topic.posts.ispublic.chronologic.active
-            unless @topic.assigned_user_id.nil?
-              agent = User.find(@topic.assigned_user_id)
-              @tracker.event(category: "Agent: #{agent.name}", action: "User Replied", label: @topic.to_param) #TODO: Need minutes
-            end
-          end
-
+          logger.info(@topic.doc_id)
+          redirect_to @topic.doc.nil? ? topic_posts_path(@topic.id) : doc_path(@topic.doc_id)
         }
       else
         format.html { render :action => "new" }
@@ -128,29 +69,7 @@ class PostsController < ApplicationController
     end
   end
 
-  def update
-    @post = Post.find(params[:id])
-    @post.body = params[:post][:body]
-    @post.active = params[:post][:active]
-
-    respond_to do |format|
-      if @post.save
-        format.js
-      end
-    end
-  end
-
-  def destroy
-    @post = Post.find(params[:id])
-    @post.destroy
-
-    respond_to do |format|
-       format.html { redirect_to topic_posts_path(@post.topic) }
-    end
-  end
-
   def up_vote
-
     if user_signed_in?
       @post = Post.find(params[:id])
       @post.votes.create(user_id: current_user.id)
@@ -162,29 +81,9 @@ class PostsController < ApplicationController
     respond_to do |format|
       format.js
     end
-
   end
-
 
   protected
-
-  def send_message
-    #Should only send when admin posts, not when user replies
-
-    if current_user.admin?
-      if @post.kind == 'first'
-        email_locale = I18n.locale
-      else
-        email_locale = @topic.locale.nil? ? I18n.locale : @topic.locale.to_sym
-      end
-
-      I18n.with_locale(email_locale) do
-        TopicMailer.new_ticket(@post.topic).deliver_later if @topic.private?
-      end
-    else
-      logger.info("reply is not from admin, don't email") #might want to cchange this if we want admin notification emails
-    end
-  end
 
   def view_causes_vote
     if logged_in?
