+++ app/views/admin/posts/cancel.js.erb
