+++ app/views/admin/posts/create.js.erb
