+++ app/views/admin/posts/edit.js.erb
@@ -1,0 +1,1 @@
+$('#post-body-<%= @post.id %>').html("<%= j render 'admin/posts/edit' %>");
