+++ app/views/admin/posts/update.js.erb
@@ -1,0 +1,1 @@
+$('#post-body-<%= @post.id %>').html("<%= j simple_format(@post.body) %>");
