+++ app/views/admin/topics/_post.html.erb
@@ -8,7 +8,7 @@
 		<span id="row-<%= post.id %>"  class="post-menu post-menu-<%= post.id %> btn-group">
 			<%= post_message(post,true) %>
 			<ul class="dropdown-menu dropdown-menu-left" role="menu">
-				<li><%= link_to 'Edit', edit_post_path(post), remote: true %></li>
+				<li><%= link_to 'Edit', edit_admin_post_path(post), remote: true %></li>
 				<li><%= link_to 'Turn post into content', new_admin_doc_path(post_id: post.id, lang: I18n.locale) %></li>
 			</ul>
 		</span>
