+++ app/views/admin/topics/_ticket.html.erb
@@ -24,7 +24,7 @@
 
       <%#= error_messages_for :post -%>
 
-      <%= bootstrap_form_for @topic.posts.new, :url => topic_posts_path(@topic), :remote => true do |f| -%>
+      <%= bootstrap_form_for @topic.posts.new, url: admin_topic_posts_path(@topic), remote: true do |f| %>
         <%= hidden_field_tag :authenticity_token, form_authenticity_token -%>
 
         <%= f.form_group :kind  do %>
