--- app/views/posts/edit.js.erb
@@ -1,1 +1,0 @@
-$('#post-body-<%= @post.id %>').html("<%= j render 'posts/edit' %>");
