--- app/views/posts/update.js.erb
@@ -1,2 +1,0 @@
-//$('#post-<%= @post.id %>').replace("<%= j render partial: 'posts/post', locals: {post: @post} %>");
-$('#post-body-<%= @post.id %>').html("<%= j simple_format(@post.body) %>");
