+++ test/controllers/posts_controller_test.rb
@@ -68,14 +68,6 @@ class PostsControllerTest < ActionController::TestCase
     assert :success
   end
 
-  test "a signed in user should NOT be able to edit a post" do
-    sign_in users(:user)
-    original_post = Post.find(1)
-    xhr :patch, :update, { id: 1, post: { body: "this has changed" }, locale: :en}
-    assert original_post.body == Post.find(1).body
-    assert :success
-  end
-
   test "a signed in user should be able to post a reply to an open topic and the topic status should change to pending" do
     sign_in users(:user)
     assert_difference "Topic.where(current_status: 'pending').count", 1 do
@@ -87,52 +79,6 @@ class PostsControllerTest < ActionController::TestCase
     sign_in users(:user)
     assert_difference "Post.find(6).points", 1 do
       xhr :post, :up_vote, id: 6, topic_id: 5, locale: :en
-    end
-  end
-
-
-  # logged in admin
-
-  # TODO this belongs in the admin controller test where all admin logic resides
-  test "an admin should be able to see inactive posts" do
-
-  end
-
-
-  test "an admin should be able to reply to a private topic, and the system should send an email" do
-    sign_in users(:admin)
-
-    assert_difference "ActionMailer::Base.deliveries.size", 1 do
-      assert_difference "Post.count", 1 do
-        xhr :post, :create, topic_id: 1, locale: :en, post: { user_id: User.find(2).id, body: "new reply", kind: "reply" }
-      end
-    end
-    assert :success
-  end
-
-  test "an admin should be able to reply to a public topic, and the system should NOT send an email" do
-    sign_in users(:admin)
-
-    assert_difference "ActionMailer::Base.deliveries.size", 0 do
-      assert_difference "Post.count", 1 do
-        xhr :post, :create, topic_id: 4, locale: :en, post: { user_id: User.find(2).id, body: "new reply", kind: "reply" }
-      end
-    end
-    assert :success
-  end
-
-  test "an admin should be able to edit a post" do
-    sign_in users(:admin)
-    old = Post.find(1).body
-    xhr :patch, :update, {id: 1, locale: :en, post: { body: "this has changed" }  }
-    assert old != Post.find(1).body
-    assert :success
-  end
-
-  test "an admin should be able to post a reply to a pending ticket, and the topic status should change to open" do
-    sign_in users(:admin)
-    assert_difference "Topic.where(current_status: 'open').count", 1 do
-      xhr :post, :create, topic_id: 2, locale: :en, post: { user_id: User.find(1).id, body: "new reply", kind: "reply" }
     end
   end
